models/evaluate/requirements.txt
toml
insightface
onnxruntime-gpu  # Jeśli nie masz GPU NVIDIA, zmień to na "onnxruntime"
faiss-gpu        # Jeśli nie masz GPU NVIDIA, zmień to na "faiss-cpu"
google-cloud-storage
numpy
opencv-python-headless
tqdm
scikit-learn
matplotlib